+++ feature/explore/src/main/java/com/teamoffroad/feature/explore/presentation/component/ExploreSuccessDialogItem.kt
@@ -1,0 +1,21 @@
+package com.teamoffroad.feature.explore.presentation.component
+
+import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import coil.compose.AsyncImage
+
+@Composable
+fun ExploreSuccessDialogContent(
+    url: String,
+    modifier: Modifier = Modifier,
+) {
+    AsyncImage(
+        model = url,
+        contentDescription = null,
+        modifier = modifier
+            .wrapContentSize(),
+        contentScale = ContentScale.FillHeight
+    )
+}
